Move WorldSettingsModal to separate component + add EventBus
